Move factory aspect to ActivationFactory and leave AgentToolRegistry as a meeting of the 2 main implementations
