+++ neuro_san/internals/graph/interfaces/agent_tool_factory.py
@@ -13,13 +13,12 @@ from typing import Any
 from typing import Dict
 
 from neuro_san.internals.graph.interfaces.callable_activation import CallableActivation
-from neuro_san.internals.run_context.interfaces.agent_network_inspector import AgentNetworkInspector
 from neuro_san.internals.run_context.interfaces.run_context import RunContext
 
 
-class AgentToolFactory(AgentNetworkInspector):
+class AgentToolFactory:
     """
-    AgentNetworkInspector Interface extension describing a factory that creates agent tools.
+    Interface describing a factory that creates agent tools.
     Having this interface breaks some circular dependencies.
     """
 
@@ -46,5 +45,18 @@ class AgentToolFactory(AgentNetworkInspector):
     def get_agent_tool_path(self) -> str:
         """
         :return: The path under which tools for this registry should be looked for.
+        """
+        raise NotImplementedError
+
+    def get_config(self) -> Dict[str, Any]:
+        """
+        :return: The entire config dictionary given to the instance.
+        """
+        raise NotImplementedError
+
+    def get_name_from_spec(self, agent_spec: Dict[str, Any]) -> str:
+        """
+        :param agent_spec: A single agent to register
+        :return: The agent name as per the spec
         """
         raise NotImplementedError
